rocky-engine/requirements.txt
fastapi
uvicorn
pydantic
groq
python-dotenv
edge-tts
SpeechRecognition
